+++ src/lib.rs
@@ -171,6 +171,9 @@
 //! [`ServiceStatus::checkpoint`]: service::ServiceStatus::checkpoint
 //! [`StartPending`]: service::ServiceState::StartPending
 //! [`Running`]: service::ServiceState::Running
+//! [`ServiceActionType`]: service::ServiceActionType
+//! [`ServiceErrorControl`]: service::ServiceErrorControl
+//! [`ServiceState`]: service::ServiceState
 
 #![cfg(windows)]
 
@@ -229,6 +232,18 @@ pub enum Error {
     /// Invalid raw representation of [`ServiceErrorControl`].
     #[error(display = "Invalid service error control value")]
     InvalidServiceErrorControl(#[error(cause)] service::ParseRawError),
+
+    /// Invalid raw representation of [`ServiceActionType`]
+    #[error(display = "Invalid service action type")]
+    InvalidServiceActionType(#[error(cause)] service::ParseRawError),
+
+    /// Invalid reboot message
+    #[error(display = "Invalid service action failures reboot message")]
+    InvalidServiceActionFailuresRebootMessage(#[error(cause)] widestring::NulError),
+
+    /// Invalid command
+    #[error(display = "Invalid service action failures command")]
+    InvalidServiceActionFailuresCommand(#[error(cause)] widestring::NulError),
 
     /// IO error when calling winapi
     #[error(display = "IO error in winapi call")]